modules/kvs: Re-org commit functions
Move commit_process(), store_cache(), commit_unroll(), and
commit_link_dirent() into commit API.  Latter 3 can be made
static and hidden from caller.
