+++ src/modules/kvs/kvs.c
@@ -85,13 +85,6 @@ typedef struct {
     const char *hash_name;
 } kvs_ctx_t;
 
-typedef enum {
-    COMMIT_PROCESS_ERROR = 1,
-    COMMIT_PROCESS_LOAD_MISSING_REFS = 2,
-    COMMIT_PROCESS_DIRTY_CACHE_ENTRIES = 3,
-    COMMIT_PROCESS_FINISHED = 4,
-} commit_process_t;
-
 static int setroot_event_send (kvs_ctx_t *ctx, json_object *names);
 static int error_event_send (kvs_ctx_t *ctx, json_object *names, int errnum);
 static void commit_prep_cb (flux_reactor_t *r, flux_watcher_t *w,
@@ -299,35 +292,6 @@ error:
     return -1;
 }
 
-/* Store object 'o' under key 'ref' in local cache. */
-static int store_cache (commit_t *c, int current_epoch, json_object *o,
-                        href_t ref, struct cache_entry **hpp)
-{
-    struct cache_entry *hp;
-    int rc = -1;
-
-    if (json_hash (c->cm->hash_name, o, ref) < 0) {
-        log_err ("json_hash");
-        goto done;
-    }
-    if (!(hp = cache_lookup (c->cm->cache, ref, current_epoch))) {
-        hp = cache_entry_create (NULL);
-        cache_insert (c->cm->cache, ref, hp);
-    }
-    if (cache_entry_get_valid (hp)) {
-        c->cm->noop_stores++;
-        json_object_put (o);
-    } else {
-        cache_entry_set_json (hp, o);
-        cache_entry_set_dirty (hp, true);
-        cache_entry_set_content_store_flag (hp, true);
-    }
-    *hpp = hp;
-    rc = 0;
- done:
-    return rc;
-}
-
 static void setroot (kvs_ctx_t *ctx, const char *rootdir, int rootseq)
 {
     if (rootseq == 0 || rootseq > ctx->rootseq) {
@@ -337,285 +301,6 @@ static void setroot (kvs_ctx_t *ctx, const char *rootdir, int rootseq)
         wait_runqueue (ctx->watchlist);
         ctx->watchlist_lastrun_epoch = ctx->epoch;
     }
-}
-
-/* Store DIRVAL objects, converting them to DIRREFs.
- * Store (large) FILEVAL objects, converting them to FILEREFs.
- * Return false and enqueue wait_t on cache object(s) if any are dirty.
- */
-static int commit_unroll (commit_t *c, int current_epoch, json_object *dir)
-{
-    json_object_iter iter;
-    json_object *subdir, *value;
-    const char *s;
-    href_t ref;
-    int rc = -1;
-    struct cache_entry *hp;
-
-    json_object_object_foreachC (dir, iter) {
-        if (json_object_object_get_ex (iter.val, "DIRVAL", &subdir)) {
-            if (commit_unroll (c, current_epoch, subdir) < 0) /* depth first */
-                goto done;
-            json_object_get (subdir);
-            if (store_cache (c, current_epoch, subdir, ref, &hp) < 0)
-                goto done;
-            if (cache_entry_get_dirty (hp)) {
-                if (zlist_push (c->dirty_cache_entries, hp) < 0)
-                    oom ();
-            }
-            json_object_object_add (dir, iter.key,
-                                    dirent_create ("DIRREF", ref));
-        }
-        else if (json_object_object_get_ex (iter.val, "FILEVAL", &value)
-                                && (s = Jtostr (value))
-                                && strlen (s) > BLOBREF_MAX_STRING_SIZE) {
-            json_object_get (value);
-            if (store_cache (c, current_epoch, value, ref, &hp) < 0)
-                goto done;
-            if (cache_entry_get_dirty (hp)) {
-                if (zlist_push (c->dirty_cache_entries, hp) < 0)
-                    oom ();
-            }
-            json_object_object_add (dir, iter.key,
-                                    dirent_create ("FILEREF", ref));
-        }
-    }
-    rc = 0;
-done:
-    return rc;
-}
-
-/* link (key, dirent) into directory 'dir'.
- */
-static int commit_link_dirent (commit_t *c, int current_epoch,
-                               json_object *rootdir, const char *key,
-                               json_object *dirent, const char **missing_ref)
-{
-    char *cpy = xstrdup (key);
-    char *next, *name = cpy;
-    json_object *dir = rootdir;
-    json_object *o, *subdir = NULL, *subdirent;
-    int rc = -1;
-
-    /* Special case root
-     */
-    if (strcmp (name, ".") == 0) {
-        errno = EINVAL;
-        goto done;
-    }
-
-    /* This is the first part of a key with multiple path components.
-     * Make sure that it is a directory in DIRVAL form, then recurse
-     * on the remaining path components.
-     */
-    while ((next = strchr (name, '.'))) {
-        *next++ = '\0';
-        if (!json_object_object_get_ex (dir, name, &subdirent)) {
-            if (!dirent) /* key deletion - it doesn't exist so return */
-                goto success;
-            if (!(subdir = json_object_new_object ()))
-                oom ();
-            json_object_object_add (dir, name, dirent_create ("DIRVAL",subdir));
-            json_object_put (subdir);
-        } else if (json_object_object_get_ex (subdirent, "DIRVAL", &o)) {
-            subdir = o;
-        } else if (json_object_object_get_ex (subdirent, "DIRREF", &o)) {
-            const char *ref = json_object_get_string (o);
-            if (!(subdir = cache_lookup_and_get_json (c->cm->cache,
-                                                      ref,
-                                                      current_epoch))) {
-                *missing_ref = ref;
-                goto success; /* stall */
-            }
-            /* do not corrupt store by modify orig. */
-            subdir = json_object_copydir (subdir);
-            json_object_object_add (dir, name, dirent_create ("DIRVAL",subdir));
-            json_object_put (subdir);
-        } else if (json_object_object_get_ex (subdirent, "LINKVAL", &o)) {
-            assert (json_object_get_type (o) == json_type_string);
-            char *nkey = xasprintf ("%s.%s", json_object_get_string (o), next);
-            if (commit_link_dirent (c,
-                                    current_epoch,
-                                    rootdir,
-                                    nkey,
-                                    dirent,
-                                    missing_ref) < 0) {
-                free (nkey);
-                goto done;
-            }
-            free (nkey);
-            goto success;
-        } else {
-            if (!dirent) /* key deletion - it doesn't exist so return */
-                goto success;
-            if (!(subdir = json_object_new_object ()))
-                oom ();
-            json_object_object_add (dir, name, dirent_create ("DIRVAL",subdir));
-            json_object_put (subdir);
-        }
-        name = next;
-        dir = subdir;
-    }
-    /* This is the final path component of the key.  Add it to the directory.
-     */
-    if (dirent)
-        json_object_object_add (dir, name, json_object_get (dirent));
-    else
-        json_object_object_del (dir, name);
-success:
-    rc = 0;
-done:
-    free (cpy);
-    return rc;
-}
-
-/* return COMMIT_PROCESS_ERROR on error,
- * COMMIT_PROCESS_LOAD_MISSING_REFS stall & load,
- * COMMIT_PROCESS_DIRTY_CACHE_ENTRIES stall & process dirty cache entries
- * COMMIT_PROCESS_FINISHED all done
- */
-static commit_process_t commit_process (commit_t *c,
-                                        int current_epoch,
-                                        const href_t rootdir_ref)
-{
-    /* Incase user calls commit_process() again */
-    if (c->errnum)
-        return COMMIT_PROCESS_ERROR;
-
-    switch (c->state) {
-        case COMMIT_STATE_INIT:
-        case COMMIT_STATE_LOAD_ROOT:
-        {
-            /* Make a copy of the root directory.
-             */
-            json_object *rootdir;
-
-            assert (!zlist_first (c->missing_refs));
-
-            c->state = COMMIT_STATE_LOAD_ROOT;
-
-            if (!(rootdir = cache_lookup_and_get_json (c->cm->cache,
-                                                       rootdir_ref,
-                                                       current_epoch))) {
-                if (zlist_push (c->missing_refs, (void *)rootdir_ref) < 0)
-                    oom ();
-                goto stall_load;
-            }
-
-            c->rootcpy = json_object_copydir (rootdir);
-
-            c->state = COMMIT_STATE_APPLY_OPS;
-            /* fallthrough */
-        }
-        case COMMIT_STATE_APPLY_OPS:
-        {
-            /* Apply each op (e.g. key = val) in sequence to the root
-             * copy.  A side effect of walking key paths is to convert
-             * DIRREFs to DIRVALs in the copy. This allows the commit
-             * to be self-contained in the rootcpy until it is
-             * unrolled later on.
-             */
-            if (fence_get_json_ops (c->f)) {
-                json_object *op, *dirent;
-                const char *missing_ref = NULL;
-                const char *key;
-                json_object *ops = fence_get_json_ops (c->f);
-                int i, len = json_object_array_length (ops);
-
-                assert (!zlist_first (c->missing_refs));
-
-                for (i = 0; i < len; i++) {
-                    missing_ref = NULL;
-                    if (!(op = json_object_array_get_idx (ops, i))
-                        || !Jget_str (op, "key", &key))
-                        continue;
-                    dirent = NULL;
-                    /* can be NULL for unlink */
-                    (void)Jget_obj (op, "dirent", &dirent);
-                    if (commit_link_dirent (c,
-                                            current_epoch,
-                                            c->rootcpy,
-                                            key,
-                                            dirent,
-                                            &missing_ref) < 0) {
-                        c->errnum = errno;
-                        break;
-                    }
-                    if (missing_ref) {
-                        if (zlist_push (c->missing_refs,
-                                        (void *)missing_ref) < 0)
-                            oom ();
-                    }
-                }
-
-                if (c->errnum != 0)
-                    return COMMIT_PROCESS_ERROR;
-
-                if (zlist_first (c->missing_refs))
-                    goto stall_load;
-            }
-            c->state = COMMIT_STATE_STORE;
-            /* fallthrough */
-        }
-        case COMMIT_STATE_STORE:
-        {
-            /* Unroll the root copy.
-             * When a DIRVAL is found, store an object and replace it
-             * with a DIRREF.  Finally, store the unrolled root copy
-             * as an object and keep its reference in c->newroot.
-             * Flushes to content cache are asynchronous but we don't
-             * proceed until they are completed.
-             */
-            struct cache_entry *hp;
-
-            assert (!zlist_first (c->dirty_cache_entries));
-
-            if (commit_unroll (c, current_epoch, c->rootcpy) < 0)
-                c->errnum = errno;
-            else if (store_cache (c,
-                                  current_epoch,
-                                  c->rootcpy,
-                                  c->newroot,
-                                  &hp) < 0)
-                c->errnum = errno;
-            else if (cache_entry_get_dirty (hp)
-                     && zlist_push (c->dirty_cache_entries, hp) < 0)
-                oom ();
-
-            if (c->errnum)
-                return COMMIT_PROCESS_ERROR;
-
-            /* cache took ownership of rootcpy, we're done, but
-             * may still need to stall user.
-             */
-            c->state = COMMIT_STATE_PRE_FINISHED;
-            c->rootcpy = NULL;
-
-            if (zlist_first (c->dirty_cache_entries))
-                goto stall_store;
-
-            /* fallthrough */
-        }
-        case COMMIT_STATE_PRE_FINISHED:
-            c->state = COMMIT_STATE_FINISHED;
-            /* fallthrough */
-        case COMMIT_STATE_FINISHED:
-            break;
-        default:
-            log_msg ("invalid commit state: %d", c->state);
-            c->errnum = EPERM;
-            return COMMIT_PROCESS_ERROR;
-    }
-
-    return COMMIT_PROCESS_FINISHED;
-
-stall_load:
-    c->blocked = 1;
-    return COMMIT_PROCESS_LOAD_MISSING_REFS;
-
-stall_store:
-    c->blocked = 1;
-    return COMMIT_PROCESS_DIRTY_CACHE_ENTRIES;
 }
 
 struct commit_cb_data {
